+++ components/JoinGroupForm.tsx
@@ -67,6 +67,9 @@ export const JoinGroupForm = ({ group }: JoinGroupFormProps): JSX.Element => {
       unmounted = true;
     };
     return cleanup;
+
+    // authUser の変更のみを検知
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [authUser]);
 
   const {
